requirements.txt
Django==1.8.2
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.9.0
Unidecode==0.04.18
Whoosh==2.7.0
argparse==1.2.1
django-datetime-widget==0.9.3
django-filer==0.9.12
django-haystack==2.4.0
django-jinja==1.4.1
django-lockdown==1.1
django-mptt==0.7.4
django-multiupload==0.5
django-polymorphic==0.7.1
django-reversion==1.8.7
django-suit==0.2.13
django-suit-redactor==0.0.2
django-taggit==0.14.0
django-taggit-autosuggest==0.2.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
easy-thumbnails==2.2
mistune==0.6
pytz==2015.4
wsgiref==0.1.2
